feat: finish splitting up parsers
