+++ crates/aiken-lang/Cargo.toml
@@ -7,9 +7,9 @@ repository = "https://github.com/aiken-lang/aiken"
 homepage = "https://github.com/aiken-lang/aiken"
 license = "Apache-2.0"
 authors = [
-    "Lucas Rosa <x@rvcas.dev>",
-    "Kasey White <kwhitemsg@gmail.com>",
-    "KtorZ <matthias.benkort@gmail.com>",
+  "Lucas Rosa <x@rvcas.dev>",
+  "Kasey White <kwhitemsg@gmail.com>",
+  "KtorZ <matthias.benkort@gmail.com>",
 ]
 rust-version = "1.66.1"
 
@@ -30,15 +30,12 @@ num-bigint = "0.4.3"
 [target.'cfg(not(target_family="wasm"))'.dependencies]
 chumsky = "0.9.2"
 [target.'cfg(target_family="wasm")'.dependencies]
-chumsky = { version = "0.9.2", features = ["ahash", "std"], default-features = false }
+chumsky = { version = "0.9.2", features = [
+  "ahash",
+  "std",
+], default-features = false }
 
 [dev-dependencies]
 indoc = "2.0.1"
-insta = { version = "1.30.0", features = ["yaml"] }
+insta.workspace = true
 pretty_assertions = "1.3.0"
-
-[profile.dev.package.insta]
-opt-level = 3
-
-[profile.dev.package.similar]
-opt-level = 3
